Create rangeFinder.js
Find the first value of the sheet/array. This will look up the first match of the specific value. This will give you the row and col that starts at 0 not 1 on the sheet/array. Make sure to add 1 when trying to use this within a getRange() formula. Great for automating within your code. No need to have a set range, this can make your range dynamic.